+++ src/main/webapp/WEB-INF/views/forgotpassword.jsp
@@ -58,45 +58,27 @@
 			<div class="wrap-login100 p-l-110 p-r-110 p-t-62 p-b-33">
 				<form class="login100-form validate-form flex-sb flex-w"
 					method="post"
-					action="${pageContext.request.contextPath }/user/login">
-					<span class="login100-form-title p-b-53"> Sign In With </span>
-
-					<!-- <a href="#" class="btn-face m-b-20">
-						<i class="fa fa-facebook-official"></i>
-						Facebook
-					</a> -->
-
-					<a style="width: 100%;" href="#" class="btn-google m-b-20"> <img
-						src=<c:url value='/template/images/icons/icon-google.png'/>
-						alt="GOOGLE"> Google
-					</a>
+					action="${pageContext.request.contextPath }/user/forgot">
+					
 
 					<div class="p-t-31 p-b-9">
-						<span class="txt1"> Username </span>
+						<span class="txt1"> Email </span>
 					</div>
 					<div class="wrap-input100 validate-input"
-						data-validate="Username is required">
-						<input class="input100" type="text" name="username"> <span
+						data-validate="Email is required">
+						<input class="input100" type="email" name="email" > <span
 							class="focus-input100"></span>
 					</div>
-
-					<div class="p-t-13 p-b-9">
-						<span class="txt1"> Password </span> <a href="${pageContext.request.contextPath }/user/forgot" id='forgot'
-							class="txt2 bo1 m-l-5"> Forgot? </a>
+					<div class="p-t-31 p-b-9">
+						<span class="txt1" style=color:"red";> ${error} </span>
 					</div>
-					<div class="wrap-input100 validate-input"
-						data-validate="Password is required">
-						<input class="input100" type="password" name="password">
-						<span class="focus-input100"></span>
-					</div>
-
 					<div class="container-login100-form-btn m-t-17">
-						<button id='login' type="submit" class="login100-form-btn">Sign In</button>
+						<button id='login' class="login100-form-btn" type="submit">Gửi Email</button>
 					</div>
-
+					
 					<div class="w-full text-center p-t-55">
-						<span class="txt2"> Not a member? </span> <a href="#"
-							class="txt2 bo1"> Sign up now </a>
+						<span class="txt2"> Quay </span> <a href="${pageContext.request.contextPath }/user/login"
+							class="txt2 bo1"> Trở lại </a>
 					</div>
 				</form>
 			</div>
