+++ src/main/webapp/index.jsp
@@ -8,6 +8,6 @@
 </head>
 <body>
 	<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
-	<c:redirect url="/user/list" />
+	<c:redirect url="/user/login" />
 </body>
 </html>
